requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
pydantic==1.10.7
pydantic-settings==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.1.0
pytesseract==0.3.10
pdf2image==1.16.3
openpyxl==3.10.0
pandas==2.0.0
openai==0.27.0
python-dotenv==1.0.0
